changes : 1. hard code is removed. 2. Now file download will not try always on the failed node and will go for next node which is alive. 3. Combiner and reducer is implemented for Sum.

+++ spark/src/main/java/com/talentica/hungryHippos/rdd/main/SumJobTextFile.java
@@ -61,7 +61,7 @@ public class SumJobTextFile implements Serializable {
         cacheRDD.put(keyOfHHRDD, hipposRDD);
       }
       Broadcast<Job> jobBroadcast = context.broadcast(job);
-      JavaRDD<Tuple2<String, Integer>> resultRDD =
+      JavaRDD<Tuple2<String, Long>> resultRDD =
           new SumExecutor<String>().process(hipposRDD, descriptionBroadcast, jobBroadcast);
       String outputDistributedPath =
           outputDirectory + File.separator + jobBroadcast.value().getJobId();
